cleanup build pipeline
